Fix SQLite decimal compatibility in LINQ queries
- CustomerService: cast TotalSpent to double for OrderByDescending
- SaleService: cast decimal fields to double in SumAsync calls
- SaleService: convert double results back to decimal after aggregation
- SaleService: cast Subtotal to double in GroupBy Sum for TopProducts
- SaleService: cast Total to double in GroupBy Sum for DailySales

+++ Services/CustomerService.cs
@@ -14,7 +14,8 @@ public class CustomerService : ICustomerService
         var q = _db.Customers.AsQueryable();
         if (!string.IsNullOrWhiteSpace(search))
             q = q.Where(c => c.Name.Contains(search) || (c.Phone != null && c.Phone.Contains(search)));
-        return q.OrderByDescending(c => c.TotalSpent).ToListAsync();
+        // SQLite 不支持直接对 decimal 排序，转为 double
+        return q.OrderByDescending(c => (double)c.TotalSpent).ToListAsync();
     }
 
     public Task<Customer?> GetByIdAsync(int id) => _db.Customers.FindAsync(id).AsTask();
